internal/executor: factor out the serving component to allow reusage of executor in different contexts
